refactor(clippy)

+++ src/utils/time.rs
@@ -15,7 +15,7 @@ pub fn get_utc_timestamp(month: u32, day: u32, hour: u32, minute: u32) -> Result
         .context("time error")?
         .and_hms_opt(hour, minute, 0)
         .context("time error")?;
-    Ok((date_time.timestamp() - 60 * 60 * 8) as i64)
+    Ok(date_time.timestamp() - 60 * 60 * 8)
 }
 
 pub fn get_now_timestamp() -> Result<i64> {
@@ -23,7 +23,7 @@ pub fn get_now_timestamp() -> Result<i64> {
     let offset: chrono::FixedOffset =
         FixedOffset::east_opt(8 * 3600).expect("time zone offset fail");
     let time = Utc::now().with_timezone(&offset);
-    Ok(time.timestamp() as i64)
+    Ok(time.timestamp())
 }
 
 pub fn get_date_with_offset(fmt: &str, day: i64) -> String {
